db-sync-service/requirements.txt
# ==============================================================================
# DB Sync Service 依賴套件
# ==============================================================================

# 核心依賴 - PostgreSQL 非同步驅動
asyncpg==0.29.0

# 環境變數管理
python-dotenv==1.0.1

# 日期時間處理
python-dateutil==2.8.2

# 可選依賴 - 健康檢查端點（如需要）
# fastapi==0.104.1
# uvicorn==0.24.0
